refactor: created the controllers for category, order and product
