+++ server/controllers/order.controller.js
@@ -1,0 +1,141 @@
+const Order = require('../models/order.model')
+const OrderItem = require('../models/order-item.model')
+
+const getOrderList = async (req, res) => {
+    const orderList = await Order.find()
+        .populate('user', 'name')
+        .sort({ createdAt: -1 })
+
+    if (!orderList) {
+        res.status(500).json({ success: false })
+    }
+    res.send(orderList)
+}
+
+const getOrderById = async (req, res) => {
+    const order = await Order.findById(req.params.id)
+        .populate('user', 'name')
+        // .populate('orderItems')
+        // populating the product inside the array of orderItems instead
+        .populate({
+            path: 'orderItems', populate: { 
+                path: 'product', populate: 'category' },
+        })
+
+    if (!order) {
+        res.status(500).json({ success: false })
+    }
+    res.send(order)
+}
+
+const postOrder = async (req, res) => {
+    const orderItemsIds = Promise.all(
+        req.body.orderItems.map(async (orderItem) => {
+            let newOrderItem = new OrderItem({
+                quantity: orderItem.quantity,
+                product: orderItem.product,
+            })
+
+            newOrderItem = await newOrderItem.save()
+
+            return newOrderItem._id
+        })
+    )
+    const orderItemsIdsResolved = await orderItemsIds
+
+    const totalPrices = await Promise.all(
+        orderItemsIdsResolved.map(async (orderItemId) => {
+            const orderItem = await OrderItem.findById(orderItemId).populate(
+                'product',
+                'price'
+            )
+            const totalPrice = orderItem.product.price * orderItem.quantity
+            return totalPrice
+        })
+        )
+        // console.log(totalPrices)
+
+    const totalPrice = totalPrices.reduce((a, b) => a + b, 0)
+
+    let order = new Order({
+        orderItems: orderItemsIdsResolved,
+        shippingAddress1: req.body.shippingAddress1,
+        shippingAddress2: req.body.shippingAddress2,
+        city: req.body.city,
+        zip: req.body.zip,
+        country: req.body.country,
+        phone: req.body.phone,
+        status: req.body.status,
+        totalPrice: totalPrice,
+        user: req.body.user,
+    })
+    order = await order.save()
+
+    if (!order) return res.status(400).send('the order cannot be created!')
+
+    res.send(order)
+}
+
+const updateOrder = async (req, res)=> {
+    const order = await Order.findByIdAndUpdate(
+        req.params.id,
+        {
+            status: req.body.status
+        },
+        { new: true}
+    )
+
+    if(!order)
+    return res.status(400).send('the order cannot be update!')
+
+    res.send(order);
+}
+
+const deleteOrder = (req, res)=>{
+    Order.findByIdAndRemove(req.params.id).then(async order =>{
+        if(order) {
+            await order.orderItems.map(async orderItem => {
+                await OrderItem.findByIdAndRemove(orderItem)
+            })
+            return res.status(200).json({success: true, message: 'the order is deleted!'})
+        } else {
+            return res.status(404).json({success: false , message: "order not found!"})
+        }
+    }).catch(err=>{
+       return res.status(500).json({success: false, error: err})
+    })
+}
+
+const getTotalSales = async (req, res)=> {
+    const totalSales= await Order.aggregate([
+        { $group: { _id: null , totalsales : { $sum : '$totalPrice'}}}
+    ])
+
+    if(!totalSales) {
+        return res.status(400).send('The order sales cannot be generated')
+    }
+
+    res.send({totalsales: totalSales.pop().totalsales})
+}
+
+const getUserOrderList = async (req, res) =>{
+    const userOrderList = await Order.find({user: req.params.userid}).populate({
+        path: 'orderItems', populate: {
+            path : 'product', populate: 'category'}
+        }).sort({'createdAt': -1});
+
+    if(!userOrderList) {
+        res.status(500).json({success: false})
+    }
+    res.send(userOrderList);
+}
+
+module.exports = {
+    getOrderList,
+    getOrderById,
+    postOrder,
+    updateOrder,
+    deleteOrder,
+    getTotalSales,
+    getUserOrderList,
+}
