+++ server/routes/order.route.js
@@ -1,7 +1,8 @@
-const Order = require('../models/order.model')
+// const Order = require('../models/order.model')
+// const OrderItem = require('../models/order-item.model')
 const express = require('express')
-const OrderItem = require('../models/order-item.model')
 const orderRouter = express.Router()
+const orderController = require('../controllers/order.controller')
 
 // orderRouter.get(`/`, async (req, res) =>{
 //     const orderList = await Order.find().populate('user', 'name').sort({'dateOrdered': -1});
@@ -26,123 +27,19 @@ const orderRouter = express.Router()
 //     res.send(order);
 // })
 
-orderRouter.get(`/`, async (req, res) => {
-    const orderList = await Order.find()
-        .populate('user', 'name')
-        .sort({ createdAt: -1 })
+orderRouter.get(`/`, orderController.getOrderList)
 
-    if (!orderList) {
-        res.status(500).json({ success: false })
-    }
-    res.send(orderList)
-})
+orderRouter.get(`/:id`, orderController.getOrderById)
 
-orderRouter.get(`/:id`, async (req, res) => {
-    const order = await Order.findById(req.params.id)
-        .populate('user', 'name')
-        // .populate('orderItems')
-        // populating the product inside the array of orderItems instead
-        .populate({
-            path: 'orderItems', populate: { 
-                path: 'product', populate: 'category' },
-        })
+orderRouter.post('/', orderController.postOrder)
 
-    if (!order) {
-        res.status(500).json({ success: false })
-    }
-    res.send(order)
-})
+orderRouter.put('/:id', orderController.updateOrder)
 
-orderRouter.post('/', async (req, res) => {
-    const orderItemsIds = Promise.all(
-        req.body.orderItems.map(async (orderItem) => {
-            let newOrderItem = new OrderItem({
-                quantity: orderItem.quantity,
-                product: orderItem.product,
-            })
+orderRouter.delete('/:id', orderController.deleteOrder)
 
-            newOrderItem = await newOrderItem.save()
+orderRouter.get('/get/totalsales', orderController.getTotalSales)
 
-            return newOrderItem._id
-        })
-    )
-    const orderItemsIdsResolved = await orderItemsIds
-
-    const totalPrices = await Promise.all(
-        orderItemsIdsResolved.map(async (orderItemId) => {
-            const orderItem = await OrderItem.findById(orderItemId).populate(
-                'product',
-                'price'
-            )
-            const totalPrice = orderItem.product.price * orderItem.quantity
-            return totalPrice
-        })
-        )
-        // console.log(totalPrices)
-
-    const totalPrice = totalPrices.reduce((a, b) => a + b, 0)
-
-    let order = new Order({
-        orderItems: orderItemsIdsResolved,
-        shippingAddress1: req.body.shippingAddress1,
-        shippingAddress2: req.body.shippingAddress2,
-        city: req.body.city,
-        zip: req.body.zip,
-        country: req.body.country,
-        phone: req.body.phone,
-        status: req.body.status,
-        totalPrice: totalPrice,
-        user: req.body.user,
-    })
-    order = await order.save()
-
-    if (!order) return res.status(400).send('the order cannot be created!')
-
-    res.send(order)
-})
-
-orderRouter.put('/:id',async (req, res)=> {
-    const order = await Order.findByIdAndUpdate(
-        req.params.id,
-        {
-            status: req.body.status
-        },
-        { new: true}
-    )
-
-    if(!order)
-    return res.status(400).send('the order cannot be update!')
-
-    res.send(order);
-})
-
-orderRouter.delete('/:id', (req, res)=>{
-    Order.findByIdAndRemove(req.params.id).then(async order =>{
-        if(order) {
-            await order.orderItems.map(async orderItem => {
-                await OrderItem.findByIdAndRemove(orderItem)
-            })
-            return res.status(200).json({success: true, message: 'the order is deleted!'})
-        } else {
-            return res.status(404).json({success: false , message: "order not found!"})
-        }
-    }).catch(err=>{
-       return res.status(500).json({success: false, error: err})
-    })
-})
-
-orderRouter.get('/get/totalsales', async (req, res)=> {
-    const totalSales= await Order.aggregate([
-        { $group: { _id: null , totalsales : { $sum : '$totalPrice'}}}
-    ])
-
-    if(!totalSales) {
-        return res.status(400).send('The order sales cannot be generated')
-    }
-
-    res.send({totalsales: totalSales.pop().totalsales})
-})
-
+orderRouter.get(`/get/userorders/:userid`, orderController.getUserOrderList)
 // orderRouter.get(`/get/count`, async (req, res) =>{
 //     const orderCount = await Order.countDocuments((count) => count)
 
@@ -154,16 +51,5 @@ orderRouter.get('/get/totalsales', async (req, res)=> {
 //     });
 // })
 
-orderRouter.get(`/get/userorders/:userid`, async (req, res) =>{
-    const userOrderList = await Order.find({user: req.params.userid}).populate({
-        path: 'orderItems', populate: {
-            path : 'product', populate: 'category'}
-        }).sort({'createdAt': -1});
-
-    if(!userOrderList) {
-        res.status(500).json({success: false})
-    }
-    res.send(userOrderList);
-})
 
 module.exports = orderRouter
